requirements.txt
chromadb
groq
pandas
plotly
PyPDF2
python-dotenv
requests
sentence-transformers
streamlit